Refactor: seperate component's in Login.tsx
