+++ src/pages/Login/Login.module.scss
@@ -10,87 +10,6 @@
     justify-content: center;
   }
 
-  .loginForm {
-    position: relative;
-
-    label {
-      display: block;
-
-      @include typography(tabLabel);
-      color: $neutral900;
-    }
-
-    input {
-      width: 100%;
-      margin-top: 6px;
-      padding: 16px;
-      border: 1px solid $neutral200;
-      border-radius: 8px;
-
-      @include typography(bodyLarge);
-      color: $neutral900;
-
-      &::placeholder {
-        color: $neutral400;
-      }
-    }
-
-    small {
-      width: 100%;
-      position: absolute;
-      bottom: 65px;
-      left: 0;
-      text-align: center;
-
-      @include typography(captionSmall);
-      color: $danger300;
-    }
-
-    &__email,
-    &__password {
-      position: relative;
-
-      .removeBtn {
-        display: none;
-        position: absolute;
-        top: 47px;
-        right: 14px;
-        cursor: default;
-      }
-
-      &:focus-within {
-        input {
-          outline: none;
-          border-width: 1px;
-          border-color: $primary300;
-        }
-
-        .removeBtn {
-          display: block;
-        }
-      }
-    }
-
-    &__password {
-      margin-top: 24px;
-    }
-
-    &__submitBtn {
-      width: 100%;
-      padding: 13px 0;
-      background: $primary300;
-      margin-top: 56px;
-      border-radius: 16px;
-
-      @include typography(button);
-      color: $neutral0;
-
-      &:hover {
-        background: $primary400;
-      }
-    }
-  }
-
   .aboutAuth {
     display: flex;
     justify-content: center;
